+++ src/main/java/duke/task/Todo.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 /**
  * This class encapsulates a Todo.
  *
