+++ src/main/java/duke/util/TaskHandler.java
@@ -1,5 +1,8 @@
+package duke.util;
+
+import duke.task.*;
+import duke.exception.*;
 import java.util.ArrayList;
-import java.util.List;
 
 /**
  * This class encapsulates a Task Handler that manages all the tasks Duke can handle.
@@ -13,17 +16,14 @@ public class TaskHandler {
     private static final String TASK_ADDED_MESSAGE = "Voila! ^_^ I've added this task:\n\t  %s\n\tYou currently have %d task(s) in the list.";
     private static final String TASK_DONE_MESSAGE = "Good Job! :D I've marked this task as done:\n\t  %s\n\tYou currently have %d undone task(s) in the list.";
     private static final String TASK_DELETED_MESSAGE = "Voila! ^_^ I've deleted this task:\n\t  %s\n\tYou currently have %d task(s) in the list.";
-    private static final String TASK_LIST = "Here are the task(s) in your list! n_n\n\t";
     private static final String NO_SUCH_TASK_ID = "Awwww, I can't seem to find this task index. Try again? U_U";
     private static final String NEGATIVE_TASK_ID = "Please enter a number starting from 1! V_V";
 
-    private ArrayList<Task> taskList;
-    private final Storage storage;
+    private final ArrayList<Task> taskList;
 
     /** Instantiates a new Task Handler **/
-    public TaskHandler(ArrayList<Task> taskList, Storage storage) {
+    public TaskHandler(ArrayList<Task> taskList) {
         this.taskList = taskList;
-        this.storage = storage;
     }
 
     /**
@@ -101,9 +101,9 @@ public class TaskHandler {
     }
 
     public void updateData() throws DukeException {
-        storage.updateData(taskList);
+        Storage.updateData(taskList);
     }
- 
+
     public void printTasks() {
         Ui.printTaskList(taskList);
     }
